refactor tests a bit. break up large file.
