+++ test/NeighborSearchProcessorTestFixture.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <gtest/gtest.h>
 #include <mpi.h>
